Don't loose a selection after filtering
Given a list like this

```
  Select an item:
    Apples
    Bananas
    Grapes
  ‣ Pineapples
```

From `Pineapples`, if you type `g` the list woulf only show `Grapes` but
nothing would be selected.

I think this is counter-intuitive as I would expect if there is only one
thing in the list then pressing enter would select it, not a return a
blank string.