Add "^L" to clear the screen
I use ^L quite a lot in bash!
Half way through typing a command, I'll realise I want to clear the
screen and press ^L, this feature would be especially nice for this tool
because it may produce a much longer list than originally expected
(depending on the input) and clearing the top half of the screen would
be handy.