Fix #1: Handle when cursor is at the bottom
When the cursor is at the bottom of the terminal, the lines get printed
over the top of previously entered lines when it should have pushed
everything up and continued.

Steps to reproduce
------------------

- Open a terminal
- Make sure the cursor is towards the bottom of the window
- Run `ls | bin/pecorb | xargs car`

Solution
--------

Using `save_pos` won't work, because it relies on the row and column of
the cursor. When a new line is added, what was row 20 is now row 19.
Instead of using that feature, we just have to manually move the cursor.
`save_pos` will be useful when we can be sure no new lines will be
added (updates for example).